requirements.txt
Flask
pydub
openai-whisper

